--- packages/main/src/themes/DurationPicker.css
@@ -1,29 +1,0 @@
-:host(:not([hidden])) {
-	display: inline-block;
-}
-
-:host {
-	color: var(--sapField_TextColor);
-	background-color: var(--sapField_Background);
-}
-
-:host .ui5-duration-picker-input {
-	width: 100%;
-	color: inherit;
-	background-color: inherit;
-}
-
-.ui5-duration-picker-input-icon-button:hover {
-	cursor: pointer;
-	background: var(--sapButton_Hover_Background);
-}
-
-.ui5-duration-picker-input-icon-button:active {
-	background-color: var(--sapButton_Active_Background);
-	color: var(--sapButton_Active_TextColor);
-}
-
-.ui5-duration-picker-input-icon-button[pressed] {
-	background-color: var(--sapButton_Active_Background);
-	color: var(--sapButton_Active_TextColor);
-}
