--- packages/main/src/themes/DurationPickerPopover.css
@@ -1,29 +1,0 @@
-.ui5-duration-picker-sliders-container {
-	display: flex;
-	justify-content: center;
-	align-items: stretch;
-	direction: ltr;
-	padding: 0.5rem;
-	min-width: 18rem;
-}
-
-.ui5-duration-picker-sliders-container.ui5-phone{
-	height: 90vh;
-}
-
-.ui5-duration-picker-wheelslider {
-	padding-left: 0.25rem;
-	padding-right: 0.25rem;
-}
-
-.ui5-duration-picker-footer {
-	height: fit-content;
-	display: flex;
-	justify-content: flex-end;
-	width: 100%;
-}
-
-.ui5-duration-picker-footer > [ui5-button] {
-	margin: 1%;
-	min-width: 20%;
-}
